moved classes around and removed delta-storage module
delta-storage was added because of hadoop version conflicts
between the app and CDAP unit-test module. Those have since been
resolved, so moved storage classes back into delta-app, and moved
a couple mock classes for testing into the new delta-test module.
