+++ delta-app/pom.xml
@@ -65,11 +65,6 @@
     </dependency>
     <dependency>
       <groupId>io.cdap.delta</groupId>
-      <artifactId>delta-storage</artifactId>
-      <version>${project.version}</version>
-    </dependency>
-    <dependency>
-      <groupId>io.cdap.delta</groupId>
       <artifactId>delta-test</artifactId>
       <version>${project.version}</version>
       <scope>test</scope>
