+++ delta-app/src/main/java/io/cdap/delta/store/PropertyEvaluator.java
