+++ delta-app/src/test/java/io/cdap/delta/store/DraftServiceTest.java
@@ -40,6 +40,9 @@ import io.cdap.delta.proto.DeltaConfig;
 import io.cdap.delta.proto.DraftRequest;
 import io.cdap.delta.proto.Plugin;
 import io.cdap.delta.proto.Stage;
+import io.cdap.delta.test.mock.MockConfigurer;
+import io.cdap.delta.test.mock.MockTableAssessor;
+import io.cdap.delta.test.mock.MockTableRegistry;
 import org.junit.After;
 import org.junit.Assert;
 import org.junit.Before;
