+++ delta-app/src/test/java/io/cdap/delta/store/MockPropertyEvaluator.java
