+++ pom.xml
@@ -84,7 +84,6 @@
     <module>delta-api</module>
     <module>delta-app</module>
     <module>delta-proto</module>
-    <module>delta-storage</module>
     <module>delta-test</module>
   </modules>
